v1.4 Library and Samples without new Shields
This is an update to the Netduino Library and Samples, but does not have
the newer shields implemented yet.  I'm doing this to allow those that
are working on a project to continue.  The new shield support will be
coming soon.   Note: There will probably be additional changes as well,
but right now, this seems to be working ok.