v1,4 (no IShieldChild) Library and Samples
Same as before, but instead of the IShieldChild implementation, using a
typical virtual/override function for processData().  was tinkering with
a way to instantiate shield classes on-the-fly as-needed.  But haven't
quite worked out the details yet and the IShieldChild interface wasn't
turning out the way I wanted it, so this is what it should have been to
begin with.  Pattern Shield and Internet Shield will be here eventually.