v1.4 Pattern Shield support has been added.
v1.4 Pattern Shield support has been added to the Netduino library.   In
addition, a SimplePATTERN project has been provided as an example.
Then Internet Shield is in progress (there are some very sparse classes
already created for it, but are not very useful yet.   The Pattern
Shield, however, is now fully supported.