v1.4 (support for using less memory)
This update adds #defines to the OneSheeldMain.cs module (and checks for
them) to allow the user to specify which shields are to be instantiated.
This will allow for less memory use by commenting out those #defines for
shields that are not going to be used.  This is similar to the
IncludeShieldsDefines.h file for Arduinos.