v1.4 (make OneSheeldUser._OneSheeld static)
The more I thought about it (and the more I tested), it really is best
if _OneSheeld is static.  If you decide to derive additional modules
from the OneSheeldUser class you don't want to create another instance
of the OneSheeldClass for each module.  This will take care of that
scenario (so that that the OneSheeldClass is only created once).