v1.4 (some minor updates)
Sometimes OneSheeld.waitForAppConnection() can be a little finicky ...
so, for testing, I replace it with a delay long enough that I can get
the 1Sheeld App setup and ready ... 1 change here is to the
AdvancedInternet/Internet.cs module ... there's now a commented out
delay ... if you find waitForAppConnection never exits, give the delay a
try instead.   Also, I've found erasing the Netduino memory and
redeploying sometimes helps get the waitForAppConnection() to work ...
not sure why that would be.

The other change is just changing a literal value to a predefined
constant (in the C# specific millis() function in the OneSheeldClass.cs
module).