v1.4 (more update, cleanup)
This update adds two default values to the sendPacket method parameters
so that those parameters don't have to be used by shield modules, etc.,
unless needed.

Also, the IHttpResponseJsonResponseCallback interface has been renamed
to IHttpJsonResponseCallback to bring it in line with the
IHttpJsonArrayLengthResponseCallback interface ... so, if your sketch
uses the interface (like the AdvancedINTERNET sample does), it will need
to be renamed. [just remove the Response next to the Http and you should
be good-to-go].

The JsonKeyChain.cs module also adds some debugging code in it that
might be useful to you if you use JsonKeyChains, so I just left it in,
it's commented out at this point, to use it, just uncomment it.