v1.5
This update brings the Netduino library and samples in-line with the
v1.5 Arduino library and samples.  The new Color Detector and NFC
Shields are now supported.  Samples for both of the new Shields are
included.  Remote Shield support has been removed (just like the Arduino
library).   The Clock Shield has been updated to v1.5 equivalents as
well. The library and Sample version properties have been updated to
1.5.