v1.5 (more ColorShield updates)
SimpleCOLOR sample now uses the ISelectedCallback interface (for the
non-callback and callback examples).  Since a reconnection is required
because of the reset during a deploy, this just guarantees that the the
Palette gets set because it waits for the Shield to be active and
selected in the 1Sheeld App.   Also tweaked the ColorShield itself a
bit.  Just didn't like how memory handling was implemented (also added a
setColor function to ColorClass to allow the color to be modified after
instantiation.