v1.5 (AdvancedInternet Sample Update)
This update just adds a ISelectedCallback to the AdvancedINTERNET
sample.  and moves the INTERNET shield initialization to a user-defined
OnSelection() callback function.  This helps quite a bit during the
startup sequence (since when a new (or updated) program is uploaded (and
the Netduino is rebooted) the 1Sheeld gets reset and requires the
Application to be reconnected.