v1.5 (A Minor Serial Port Update)
Even though I don't think it will make much difference, I did find an
issue with the recent Serial Port improvements.  Basically, there is a
new function that I had created that passes in the serial port and in
one place I was using a global reference to the serial port rather than
what was passed in to the function ... again, I believe they both point
to the same place, but just to make sure, that is now fixed.  Also, a
little general cleanup and a bit of additional comments have been added.