v1.5 (Another Serial Port Handling Update)
This update adds a do-while loop to the processSerial function.  The
idea being that it is possible that additional data could come in on the
serial port while the data is being processed.  Was thinking that I'd
rather process it ASAP rather than wait for another processSerial call.
So, the while loop checks if there is additional data on the serial port
that needs processing (and if so, processes it).