v1.6 (AdvancedGLCD sample button style fix).
This is just an update to the AdvancdGLCD sample.   The way I was
handling setting the 3D button style wasn't working ... now doing it
more similarly to Arduino sample.   Now it works. (guess I can't set the
style in the Button Constructor).  No actual changes to library ... just
the Sample itself.