v1.6 (GLCD callback changes)
After talking with the guys and 1Sheeld, I realized that the GLCDSlider
callback could receive a full integral value (larger than the byte value
that was being used) ... so, the IInteractiveShapeChangeCallback.   Note
the interface has changed the OnChange method argument type to an object
and the needed object type must be cast into a proper variable (example
the GLCD Button, CheckBox and RadioButton passes a bool value where the
Slider passes an int value.   If you look at any of the AdvancedGLCD
sample button modules, you see now that the data value passed in is cast
to the proper value ( the Slider int value would just do something like
int sliderval = (int) data;  ... if there are any future interactive
GLCD controls that require other OnChange types, we'll be able to handle
them similarly, as needed, and the OnChange method argument type won't
have to change).