v1.10.0.1
Forgot to add the two Barcode Scanner samples to the main solution
(which is what I check-in to Github, but test with an individual
solution for each sample app, which I don't check-in to Github).   I did
check in the two sample projects, which means you could add them
yourself of course.